+++ include/cache.h
@@ -3,7 +3,7 @@
 
 typedef struct chash_chain_s {
 	void *ptr;
-	unsigned id, key;
+	u64 id, key;
 	struct chash_chain_s *next, *prev;
 } chash_chain_t;
 
@@ -13,7 +13,7 @@ typedef struct {
 } chash_t;
 
 struct ce_t {
-	unsigned id, key;
+	u64 id, key;
 	char dirty;
 	char *data;
 	unsigned length;
@@ -42,13 +42,13 @@ extern cache_t caches[NUM_CACHES];
 
 chash_t *chash_create(unsigned length);
 int chash_destroy(chash_t *h);
-void *chash_search(chash_t *h, unsigned id, unsigned key);
-int chash_delete(chash_t *h, unsigned id, unsigned key);
-int chash_add(chash_t *h, unsigned id, unsigned key, void *ptr);
-int destroy_all_id(cache_t *c, unsigned);
-int do_cache_object(cache_t *, unsigned id, unsigned key, int sz, char *buf, int dirty);
+void *chash_search(chash_t *h, u64 id, u64 key);
+int chash_delete(chash_t *h, u64 id, u64 key);
+int chash_add(chash_t *h, u64 id, u64 key, void *ptr);
+int destroy_all_id(cache_t *c, u64);
+int do_cache_object(cache_t *, u64 id, u64 key, int sz, char *buf, int dirty);
 cache_t * get_empty_cache(int (*)(struct ce_t *), char *);
-struct ce_t *find_cache_element(cache_t *, unsigned id, unsigned key);
+struct ce_t *find_cache_element(cache_t *, u64 id, u64 key);
 void sync_cache(cache_t *);
 int destroy_cache(cache_t *);
 int sync_element(cache_t *, struct ce_t *e);
